fix(okf): compare a heading level as a number, not against an Int
HeadingLevel replaced the bare Int on Block.Heading in 54f6a87d, which broke
this assertion:

    Values of types morphir.langkit.markdown.HeadingLevel and Int cannot be
    compared with == or !=

That commit reported morphir-knowledge-okf as still compiling, which was true
of its main sources but not of its tests -- only the former was checked. The
full test:js run caught it.

+++ morphir/knowledge/okf/test/src/morphir/knowledge/okf/ConceptTests.scala
@@ -34,7 +34,7 @@ class ConceptTests extends Test[Any]:
           assert(concept.body.blocks.size == 2)
           concept.body.blocks(0) match
             case Block.Heading(level, text, _) =>
-              assert(level == 1)
+              assert(level.toInt == 1)
               assert(text == "Title")
             case _ => assert(false)
         case _ => assert(false)
